Divided Fight trait into Defender and Attacker
It was required to clone the whole Actor before but only the attack
value was required.  Now the Attacker trait can copy only values
required for an attack into a different struct which implements the
Actor trait.  Now only a u32 is copied instead of a whole Actor
including its attributes and items.